+++ beagle/src/test/kotlin/br/com/zup/beagle/android/view/custom/BeagleViewTest.kt
@@ -25,12 +25,14 @@ import androidx.lifecycle.MutableLiveData
 import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import br.com.zup.beagle.android.BaseSoLoaderTest
 import br.com.zup.beagle.android.BaseTest
 import br.com.zup.beagle.android.MyBeagleSetup
 import br.com.zup.beagle.android.components.Text
 import br.com.zup.beagle.android.context.constant
 import br.com.zup.beagle.android.data.formatUrl
 import br.com.zup.beagle.android.networking.RequestData
+import br.com.zup.beagle.android.setup.BeagleConfigurator
 import br.com.zup.beagle.android.setup.BeagleSdk
 import br.com.zup.beagle.android.utils.ObjectWrapperForBinder
 import br.com.zup.beagle.android.view.ApplicationTest
@@ -60,7 +62,7 @@ import org.robolectric.annotation.Config
 
 @Config(application = ApplicationTest::class)
 @RunWith(AndroidJUnit4::class)
-internal class BeagleViewTest : BaseTest() {
+internal class BeagleViewTest : BaseSoLoaderTest() {
 
     @get:Rule
     var instantExecutorRule = InstantTaskExecutorRule()
