+++ beagle/src/test/kotlin/br/com/zup/beagle/android/view/viewmodel/ActionRequestViewModelTest.kt
@@ -17,6 +17,7 @@
 package br.com.zup.beagle.android.view.viewmodel
 
 import androidx.lifecycle.Observer
+import br.com.zup.beagle.android.BaseSoLoaderTest
 import br.com.zup.beagle.android.BaseTest
 import br.com.zup.beagle.android.action.SendRequestInternal
 import br.com.zup.beagle.android.data.ActionRequester
@@ -41,7 +42,7 @@ import org.junit.jupiter.api.extension.ExtendWith
 
 @ExperimentalCoroutinesApi
 @ExtendWith(InstantExecutorExtension::class, CoroutinesTestExtension::class)
-class ActionRequestViewModelTest : BaseTest() {
+class ActionRequestViewModelTest : BaseSoLoaderTest() {
 
     private val actionRequester: ActionRequester = mockk()
