+++ beagle/src/test/kotlin/br/com/zup/beagle/android/view/viewmodel/BeagleScreenViewModelTest.kt
@@ -17,6 +17,7 @@
 package br.com.zup.beagle.android.view.viewmodel
 
 import androidx.lifecycle.Observer
+import br.com.zup.beagle.android.BaseSoLoaderTest
 import br.com.zup.beagle.android.BaseTest
 import br.com.zup.beagle.android.testutil.CoroutinesTestExtension
 import br.com.zup.beagle.android.testutil.InstantExecutorExtension
@@ -32,7 +33,7 @@ import org.junit.jupiter.api.extension.ExtendWith
 
 @ExperimentalCoroutinesApi
 @ExtendWith(InstantExecutorExtension::class, CoroutinesTestExtension::class)
-class BeagleScreenViewModelTest : BaseTest() {
+class BeagleScreenViewModelTest : BaseSoLoaderTest() {
 
     private val observer: Observer<Boolean> = mockk()
